+++ imports/scenes/Contract/BillingSchedule/ProductsSection/index.js
@@ -65,6 +65,7 @@ export default class ProductsSection extends React.Component {
               name={i}
               value={charge.description}
               onChange={onChangeDescription}
+              disabled={this.props.disabled}
               type="textarea"/>
           </td>
           <td>
@@ -72,6 +73,7 @@ export default class ProductsSection extends React.Component {
               style={{textAlign: 'right'}}
               name="value"
               onChange={onChangePrice}
+              disabled={this.props.disabled}
               value={charge.value}/>
           </td>
         </tr>
@@ -95,15 +97,18 @@ export default class ProductsSection extends React.Component {
               disabled={true}
               value={moment(this.props.billingProducts[0].startDate).format("DD/MM/YYYY")}/>
             <ExpiryDate
+              disabled={this.props.disabled}
               billingProducts={this.props.billingProducts}
               updateBilling={this.props.updateBilling}
             />
             <this.props.AccountsSelector
+              disabled={this.props.disabled}
               accountsDatabase={this.props.accountsDatabase}
               charges={this.props.billingProducts}
               updateBilling={this.updateBilling}
             />
             <this.props.EqualCharges
+              disabled={this.props.disabled}
               masterValue={this.props.productsValue}
               charges={this.props.billingProducts}
               updateBilling={this.updateBilling}/>